Created command to generate full page screenshot.
