+++ venia-integration-tests/cypress.config.json
@@ -1,51 +1,49 @@
 {
-  "baseUrl": "https://develop.pwa-venia.com/",
-  "defaultCommandTimeout": 10000,
-  "supportFile": "src/support/index.js",
-  "pluginsFile": "src/plugins/index.js",
-  "fixturesFolder": "src/fixtures",
-  "integrationFolder": "src/tests",
-  "pageLoadTimeout": 30000,
-  "requestTimeout": 30000,
-  "testFiles": "**/*.spec.js",
-  "viewportHeight": 900,
-  "viewportWidth": 1440,
-  "scrollBehavior": false,
-  "ignoreTestFiles": [
-    "**/__snapshots__/*",
-    "**/__image_snapshots__/*"
-  ],
-  "env": {
-    "cypress-plugin-snapshots": {
-      "autoCleanUp": false,
-      "scrollBehavior": false,
-      "autopassNewSnapshots": true,
-      "diffLines": 3,
-      "excludeFields": [],
-      "ignoreExtraArrayItems": false,
-      "ignoreExtraFields": false,
-      "normalizeJson": true,
-      "prettier": true,
-      "imageConfig": {
-        "createDiffImage": true,
-        "resizeDevicePixelRatio": true,
-        "threshold": 0.01,
-        "thresholdType": "percent"
-      },
-      "screenshotConfig": {
-        "blackout": [],
-        "capture": "fullPage",
-        "clip": null,
-        "disableTimersAndAnimations": true,
-        "log": false,
-        "scale": false,
-        "timeout": 30000
-      },
-      "serverEnabled": true,
-      "serverHost": "localhost",
-      "serverPort": 2121,
-      "updateSnapshots": false,
-      "backgroundBlend": "difference"
+    "baseUrl": "https://develop.pwa-venia.com/",
+    "defaultCommandTimeout": 10000,
+    "supportFile": "src/support/index.js",
+    "pluginsFile": "src/plugins/index.js",
+    "fixturesFolder": "src/fixtures",
+    "integrationFolder": "src/tests",
+    "screenshotsFolder": "src/screenshots",
+    "pageLoadTimeout": 30000,
+    "requestTimeout": 30000,
+    "testFiles": "**/*.spec.js",
+    "viewportHeight": 900,
+    "viewportWidth": 1440,
+    "scrollBehavior": false,
+    "ignoreTestFiles": ["**/__snapshots__/*", "**/__image_snapshots__/*"],
+    "env": {
+        "cypress-plugin-snapshots": {
+            "autoCleanUp": false,
+            "scrollBehavior": false,
+            "autopassNewSnapshots": true,
+            "diffLines": 3,
+            "excludeFields": [],
+            "ignoreExtraArrayItems": false,
+            "ignoreExtraFields": false,
+            "normalizeJson": true,
+            "prettier": true,
+            "imageConfig": {
+                "createDiffImage": true,
+                "resizeDevicePixelRatio": true,
+                "threshold": 0.01,
+                "thresholdType": "percent"
+            },
+            "screenshotConfig": {
+                "blackout": [],
+                "capture": "fullPage",
+                "clip": null,
+                "disableTimersAndAnimations": true,
+                "log": false,
+                "scale": false,
+                "timeout": 30000
+            },
+            "serverEnabled": true,
+            "serverHost": "localhost",
+            "serverPort": 2121,
+            "updateSnapshots": false,
+            "backgroundBlend": "difference"
+        }
     }
-  }
 }
