+++ venia-integration-tests/src/support/index.js
@@ -1,1 +1,2 @@
 import 'cypress-plugin-snapshots/commands';
+import './fullPageScreenshotCommand';
